+++ clang-tools-extra/test/clang-tidy/checkers/modernize/use-scoped-lock-only-warn-on-multiple.cpp
@@ -1,5 +1,5 @@
 // RUN: %check_clang_tidy -std=c++17-or-later %s modernize-use-scoped-lock %t -- \
-// RUN:   -config="{CheckOptions: {modernize-use-scoped-lock.WarnOnlyMultipleLocks: true}}" \
+// RUN:   -config="{CheckOptions: {modernize-use-scoped-lock.WarnOnlyOnMultipleLocks: true}}" \
 // RUN:   -- -fno-delayed-template-parsing
 
 namespace std {